create applicants functions

+++ src/services/auth_service.py
@@ -3,14 +3,17 @@ from datetime import datetime
 from fastapi import HTTPException, status
 from sqlalchemy.ext.asyncio import AsyncSession
 from src.core.config import setting
-from src.cruds.user_cruds.user_crud import usercrud
-from src.cruds.user_cruds.role_crud import rolecrud
+from src.cruds import profile_crud
+from src.cruds.applicant_cruds import applicant_crud
+from src.cruds.auth_crud import authcrud
+from src.cruds.company_cruds import company_crud
+from src.cruds.role_crud import rolecrud
 from src.schemas.auth_schema import UserCreate, UserLogin, TokenResponse
 from src.redis.redis_client import RedisClient
 from src.utils.auth_utils import JWTToken
 from src.core.hash import HashService
 
-class UserService:
+class AuthService:
     def __init__(self, user_crud, role_crud):
         self.user_crud = user_crud
         self.role_crud = role_crud
@@ -23,7 +26,11 @@ class UserService:
                 detail="Пользователь с таким email уже существует"
             )
 
-        role = await self.role_crud.get_by_name(db, "applicant")
+        if user_data.role == "applicant":
+            role = await self.role_crud.get_by_name(db, "applicant")
+        else:
+            role = await self.role_crud.get_by_name(db, "company")
+
         if not role:
             raise HTTPException(
                 status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
@@ -39,19 +46,53 @@ class UserService:
             "created_at": datetime.utcnow(),
             "updated_at": datetime.utcnow()
         }
-
         user = await self.user_crud.create(db, user_dict)
+        await db.flush()
+
+        if user_data.role == "applicant":
+            applicant = await applicant_crud.applicantcrud.create(db, {})
+            await db.flush() 
+
+            profile_data = {
+                "user_id": user.id,
+                "applicant_id": applicant.id
+            }
+            await profile_crud.profilecrud.create(db, profile_data)
+
+        else: 
+            if not user_data.company_name:
+                raise HTTPException(
+                    status_code=status.HTTP_400_BAD_REQUEST,
+                    detail="Название компании обязательно для регистрации"
+                )
+            company = await company_crud.companycrud.create(db, {"name": user_data.company_name})
+            await db.flush()
+
+            profile_data = {
+                "user_id": user.id,
+                "company_id": company.id
+            }
+            await profile_crud.profilecrud.create(db, profile_data)
+
         await db.commit()
         await db.refresh(user)
         return user
 
     async def login(self, db: AsyncSession, user_data: UserLogin, redis_client: RedisClient) -> TokenResponse:
-        user = await self.user_crud.get_by_email(db, user_data.email)
+        user = await self.user_crud.get_by_email_with_role(db, user_data.email)
+        
         if not user or not HashService.verify_password(user_data.password, user.password_hash):
             raise HTTPException(status_code=401, detail="Неверный email или пароль")
+        
         if not user.is_active:
             raise HTTPException(status_code=403, detail="Пользователь деактивирован")
-
+        
+        if user.role.name != user_data.role:
+            raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail="Неверный email или пароль"
+            )
+        
         await redis_client.delete_session(str(user.id))
 
         access_jti = str(uuid.uuid4())
